Intoduce RestControllerIntegrationTestBase class
