+++ src/test/java/student/laurens/novibackend/controllers/ControllerIntegrationTestBase.java
@@ -1,0 +1,35 @@
+package student.laurens.novibackend.controllers;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.servlet.MockMvc;
+import student.laurens.novibackend.NoviBackendApplication;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = NoviBackendApplication.class)
+@AutoConfigureMockMvc
+@TestPropertySource(locations = "classpath:application-integration-test.properties")
+public abstract class ControllerIntegrationTestBase {
+
+    protected final String USER = "DefaultUser";
+    protected final String USER_ROLE = "USER";
+
+    protected final String ADMIN = "DefaultAdmin";
+    protected final String ADMIN_ROLE = "ADMIN";
+
+    @Autowired
+    protected MockMvc mvc;
+
+    public static String asJsonString(final Object obj) {
+        try {
+            return new ObjectMapper().writeValueAsString(obj);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
